Add sort declaratoins
